requirements.txt
fastapi
uvicorn
python-dotenv
beautifulsoup4
requests
playwright
sentence-transformers
faiss-cpu
groq
